Fixed "Call to undefined method getPublished" on creating folder inside object tree.

+++ lib/DivanteLtd/PimcoreElasticsearchPlugin/Event/EventManager.php
@@ -67,7 +67,7 @@ class EventManager
             return;
         }
 
-        if (!($element instanceof Asset) && !$element->getPublished()) {
+        if (!($element instanceof Asset) && method_exists($element, 'getPublished') && !$element->getPublished()) {
             $indexerService->delete($element);
             return;
         }
